Recon/requirements.txt
requests>=2.25.1
dnspython>=2.1.0
urllib3>=1.26.5 
